pyproject.toml
[tool.poetry]
authors = ["Adam Gayoso <adamgayoso@berkeley.edu>", "Galen Xing <gx2113@columbia.edu>", "Romain Lopez <romain_lopez@gmail.com>"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]
description = "Skeleton respository for generating novel scvi-tools models."
documentation = "https://scvi-tools.org"
homepage = "https://github.com/YosefLab/scvi-tools-skeleton"
license = "BSD-3-Clause"
name = "scvi-tools-skeleton"
packages = [
  {include = "scskeleton"},
]
readme = "README.rst"
version = "0.1.0"

[tool.poetry.dependencies]
anndata = ">=0.7.5"
black = {version = ">=20.8b1", optional = true}
codecov = {version = ">=2.0.8", optional = true}
flake8 = {version = ">=3.7.7", optional = true}
importlib-metadata = {version = "^1.0", python = "<3.8"}
ipython = {version = ">=7.1.1", optional = true}
jupyter = {version = ">=1.0", optional = true}
leidenalg = {version = "*", optional = true}
loompy = {version = ">=3.0.6", optional = true}
nbconvert = {version = ">=5.4.0", optional = true}
nbformat = {version = ">=4.4.0", optional = true}
nbsphinx = {version = "*", optional = true}
nbsphinx-link = {version = "*", optional = true}
pre-commit = {version = ">=2.7.1", optional = true}
prospector = {version = "*", optional = true}
pydata-sphinx-theme = {version = ">=0.4.0", optional = true}
pytest = {version = ">=4.4", optional = true}
python = ">=3.6.1,<4.0"
python-igraph = {version = "*", optional = true}
scanpy = {version = ">=1.6", optional = true}
scanpydoc = {version = ">=0.5", optional = true}
scikit-misc = {version = ">=0.1.3", optional = true}
scvi-tools = ">=0.9.0a0"
sphinx = {version = "^3.0", optional = true}
sphinx-autodoc-typehints = {version = "*", optional = true}
typing_extensions = {version = "*", python = "<3.8"}

[tool.poetry.extras]
dev = ["black", "pytest", "flake8", "codecov", "scanpy", "loompy", "jupyter", "nbformat", "nbconvert", "pre-commit", "prospector"]
docs = ["sphinx", "scanpydoc", "nbsphinx", "nbsphinx-link", "ipython", "pydata-sphinx-theme", "typing_extensions", "sphinx-autodoc-typehints"]
tutorials = ["scanpy", "leidenalg", "python-igraph", "loompy", "scikit-misc"]

[tool.poetry.dev-dependencies]

[build-system]
build-backend = "poetry.masonry.api"
requires = [
  "poetry>=1.0",
  "setuptools", # keep it here or "pip install -e" would fail
]
